fixing and tidying

+++ parallel/include/whereswally/framework/results_analysis.h
@@ -8,7 +8,7 @@
     private:
       std::vector<Pattern_Result> final_results;
     public:
-      void calculate_final_results(int number_of_final_results, std::vector<Pattern_Result> results);
+      void calculate_final_results(size_t number_of_final_results, std::vector<Pattern_Result> results);
       void add_to_final_results(std::vector<Pattern_Result> results);
       std::vector<Pattern_Result> get_final_results();
   };
